Full GL api as from https://github.com/KhronosGroup/OpenGL-Registry/blob/master/xml/gl.xml with documentation from https://github.com/KhronosGroup/OpenGL-Refpages/tree/master/gl4. Autogenerated using tool from https://github.com/devel0/netcore-sci/tree/d23959d0fba177382cd23241a6cc51a3ca3356c3/tools/gl-csharp-bindings-generator.

+++ src/Avalonia.X11/Glx/GlxDisplay.cs
@@ -132,8 +132,8 @@ namespace Avalonia.X11.Glx
 
                 var attrs = new int[]
                 {
-                    GLX_CONTEXT_MAJOR_VERSION_ARB, profile.Major,
-                    GLX_CONTEXT_MINOR_VERSION_ARB, profile.Minor,
+                    GLX_CONTEXT_MAJOR_VERSION_ARB, (int)profile.Major,
+                    GLX_CONTEXT_MINOR_VERSION_ARB, (int)profile.Minor,
                     GLX_CONTEXT_PROFILE_MASK_ARB, profileMask,
                     0
                 };
